Fix wrong preccesion

+++ src/DatabaseManipulation.java
@@ -122,7 +122,7 @@ public class DatabaseManipulation {
                 preparedStatement.setString(2, Info[1]);
                 preparedStatement.setString(3, Info[2]);
                 preparedStatement.setString(4, Info[4]);
-                preparedStatement.setFloat(5, Float.parseFloat(Info[5]));
+                preparedStatement.setDouble(5, Double.parseDouble(Info[5]));
                 preparedStatement.setDate(6, Date.valueOf(Info[6]));
                 preparedStatement.executeUpdate();
             }
@@ -190,7 +190,7 @@ public class DatabaseManipulation {
         cityR.addBatch();
 
         shipment.setString(1, ItemName);
-        shipment.setFloat(2, Float.parseFloat(ItemPrice));
+        shipment.setDouble(2, Double.parseDouble(ItemPrice));
         shipment.setString(3, ItemType);
         shipment.setString(4, RetrievalCity);
         shipment.setString(5, DeliveryCity);
@@ -209,7 +209,7 @@ public class DatabaseManipulation {
 
         courierR.setString(1, RetrievalCourier);
         courierR.setString(2, RetrievalCourierGender);
-        courierR.setDate(3, CalBirth(RetrievalStartTime, Float.parseFloat(RetrievalCourierAge)));
+        courierR.setDate(3, CalBirth(RetrievalStartTime, Double.parseDouble(RetrievalCourierAge)));
         courierR.setString(4, RetrievalCourierPhoneNumber);
         courierR.setString(5, CompanyName);
         courierR.setString(6, ItemExportCity);
@@ -227,7 +227,7 @@ public class DatabaseManipulation {
             export_detail.setString(2, "export");
             export_detail.setString(3, ItemType);
             export_detail.setString(4, ItemExportCity);
-            export_detail.setFloat(5, Float.parseFloat(ItemExportTax));
+            export_detail.setDouble(5, Double.parseDouble(ItemExportTax));
             export_detail.setDate(6, Date.valueOf(ItemExportTime));
             export_detail.addBatch();
         }
@@ -256,7 +256,7 @@ public class DatabaseManipulation {
             import_detail.setString(2, "import");
             import_detail.setString(3, ItemType);
             import_detail.setString(4, ItemImportCity);
-            import_detail.setFloat(5, Float.parseFloat(ItemImportTax));
+            import_detail.setDouble(5, Double.parseDouble(ItemImportTax));
             import_detail.setDate(6, Date.valueOf(ItemImportTime));
             import_detail.addBatch();
         }
@@ -265,7 +265,7 @@ public class DatabaseManipulation {
 
             courierD.setString(1, DeliveryCourier);
             courierD.setString(2, DeliveryCourierGender);
-            courierD.setDate(3, CalBirth(DeliveryFinishTime, Float.parseFloat(DeliveryCourierAge)));
+            courierD.setDate(3, CalBirth(DeliveryFinishTime, Double.parseDouble(DeliveryCourierAge)));
             courierD.setString(4, DeliveryCourierPhoneNumber);
             courierD.setString(5, CompanyName);
             courierD.setString(6, ItemImportCity);
@@ -345,7 +345,7 @@ public class DatabaseManipulation {
         ie_detail.executeBatch();
         shipment.executeBatch();
         long endTime = System.currentTimeMillis();
-        System.out.printf("Deleted: %d records, speed: %.2f records/s\n", name.length, (float) (name.length * 1e3 / (endTime - startTime)));
+        System.out.printf("Deleted: %d records, speed: %.2f records/s\n", name.length, (Double) (name.length * 1e3 / (endTime - startTime)));
     }
     public long deleteCity(String name) throws SQLException {
         long startTime = System.currentTimeMillis();
@@ -373,7 +373,7 @@ public class DatabaseManipulation {
         }
         city.executeBatch();
         long endTime = System.currentTimeMillis();
-        System.out.printf("Deleted City: %d records, speed: %.2f records/s\n", name.length, (float) (name.length * 1e3 / (endTime - startTime)));
+        System.out.printf("Deleted City: %d records, speed: %.2f records/s\n", name.length, (Double) (name.length * 1e3 / (endTime - startTime)));
     }
     public long deleteCourier(String name) throws SQLException {
         if (con==null) {
@@ -401,7 +401,7 @@ public class DatabaseManipulation {
         }
         courier.executeBatch();
         long endTime = System.currentTimeMillis();
-        System.out.printf("Deleted Courier: %d records, speed: %.2f records/s\n", name.length, (float) (name.length * 1e3 / (endTime - startTime)));
+        System.out.printf("Deleted Courier: %d records, speed: %.2f records/s\n", name.length, (Double) (name.length * 1e3 / (endTime - startTime)));
     }
     //Select Records
     public long selectShipmentByName(String Item) throws SQLException {
@@ -638,7 +638,7 @@ public class DatabaseManipulation {
             export_detail.setString(2, "export");
             export_detail.setString(3, ItemType);
             export_detail.setString(4, ItemExportCity);
-            export_detail.setFloat(5, Float.parseFloat(ItemExportTax));
+            export_detail.setDouble(5, Double.parseDouble(ItemExportTax));
             export_detail.setDate(6, Date.valueOf(ItemExportTime));
             export_detail.addBatch();
         }
@@ -665,7 +665,7 @@ public class DatabaseManipulation {
             import_detail.setString(2, "import");
             import_detail.setString(3, ItemType);
             import_detail.setString(4, ItemImportCity);
-            import_detail.setFloat(5, Float.parseFloat(ItemImportTax));
+            import_detail.setDouble(5, Double.parseDouble(ItemImportTax));
             import_detail.setDate(6, Date.valueOf(ItemImportTime));
             import_detail.addBatch();
         }
@@ -674,7 +674,7 @@ public class DatabaseManipulation {
 
             courierD.setString(1, DeliveryCourier);
             courierD.setString(2, DeliveryCourierGender);
-            courierD.setDate(3, CalBirth(DeliveryFinishTime, Float.parseFloat(DeliveryCourierAge)));
+            courierD.setDate(3, CalBirth(DeliveryFinishTime, Double.parseDouble(DeliveryCourierAge)));
             courierD.setString(4, DeliveryCourierPhoneNumber);
             courierD.setString(5, CompanyName);
             courierD.setString(6, ItemImportCity);
@@ -841,7 +841,7 @@ public class DatabaseManipulation {
         resultSet.next();
         System.out.println("-----------------------------------------------------------------------");
             String Port = resultSet.getString("port_city");
-            float taxrate = resultSet.getFloat("taxrate");
+            Double taxrate = resultSet.getDouble("taxrate");
 
             System.out.printf("The best port for (%s) to (%s) is (%s), tax rate is (%f)\n", Item_type, Type, Port, taxrate);
         System.out.println("-----------------------------------------------------------------------");
@@ -965,11 +965,11 @@ public class DatabaseManipulation {
     }
     return -1;
 }
-    public static Date CalBirth (String str, float age) {
+    public static Date CalBirth (String str, Double age) {
         Date date = Date.valueOf(str);
         Calendar birth = Calendar.getInstance();
         birth.setTime(date);
-        birth.add(Calendar.YEAR, -1*(int)age);
+        birth.add(Calendar.YEAR, (int)(-1*age));
         java.util.Date ret = birth.getTime();
         long temp = ret.getTime();
         Date out = new Date(temp);
